docker/dataGenerator/callCenter/requirements.txt
psycopg2-binary==2.9.9
faker==20.1.0
numpy==1.26.2
pytz~=2023.3.post1
